restrict validation to hubl files only
