+++ lib/lint.js
@@ -1,92 +1,19 @@
 const vscode = require('vscode');
-const { validateHubl } = require('@hubspot/cli-lib/api/validate');
-const { getPortalId } = require('@hubspot/cli-lib');
-const {
-  TEMPLATE_ERRORS_TYPES,
-  VSCODE_SEVERITY,
-  WORD_REGEX,
-} = require('./constants');
-const fs = require('fs');
-const path = require('path');
-
-const getRange = (document, error) => {
-  const adjustedLineNumber = error.lineno > 0 ? error.lineno - 1 : 0;
-
-  if (error.startPosition > 0) {
-    return document.getWordRangeAtPosition(
-      new vscode.Position(adjustedLineNumber, error.startPosition - 1),
-      WORD_REGEX
-    );
-  } else {
-    return document.lineAt(new vscode.Position(adjustedLineNumber, 0)).range;
-  }
-};
-
-async function updateDiagnostics(document, collection) {
-  if (!document) {
-    return collection.clear();
-  }
-
-  const { renderingErrors } = await validateHubl(
-    getPortalId(),
-    document.getText()
-  );
-
-  const isFileInWorkspace = (error) => {
-    const pathToActiveFile = vscode.window.activeTextEditor.document.uri.path;
-    const dirToActiveFile = path.dirname(pathToActiveFile);
-
-    let filePath = error.categoryErrors.fullName || error.categoryErrors.path;
-
-    if (error.category === 'MODULE_NOT_FOUND') {
-      filePath = filePath + '.module';
-    }
-
-    return fs.existsSync(path.resolve(dirToActiveFile, filePath));
-  };
-
-  const fitleredErrors = renderingErrors.filter((error) => {
-    if (
-      error.reason === TEMPLATE_ERRORS_TYPES.MISSING ||
-      error.reason === TEMPLATE_ERRORS_TYPES.BAD_URL
-    ) {
-      return isFileInWorkspace(error) ? false : true;
-    }
-    return true;
-  });
-
-  const templateErrors = fitleredErrors.map((error) => {
-    return {
-      code: '',
-      message: error.message,
-      range: getRange(document, error),
-      severity: vscode.DiagnosticSeverity[VSCODE_SEVERITY[error.reason]],
-    };
-  });
-
-  collection.set(document.uri, templateErrors);
-}
+const { triggerValidate } = require('./validation');
 
 const collection = vscode.languages.createDiagnosticCollection('hubl');
 let documentChangeListener;
 
 const enableLinting = () => {
-  let timeout = null;
-
   if (vscode.window.activeTextEditor) {
-    updateDiagnostics(vscode.window.activeTextEditor.document, collection);
+    triggerValidate(vscode.window.activeTextEditor.document, collection);
   }
 
-  documentChangeListener = vscode.workspace.onDidChangeTextDocument(
-    (editor) => {
-      if (editor) {
-        clearTimeout(timeout);
-        timeout = setTimeout(function () {
-          updateDiagnostics(editor.document, collection);
-        }, 1000);
-      }
+  documentChangeListener = vscode.workspace.onDidChangeTextDocument((event) => {
+    if (event.document) {
+      triggerValidate(event.document, collection);
     }
-  );
+  });
 };
 
 const disableLinting = () => {
